Add HighLevel damage buff

+++ tests/constants.hpp
@@ -1,7 +1,7 @@
 #ifndef CONSTANTS_HPP
 #define CONSTANTS_HPP
 
-constexpr int initial_health = 1000;
-constexpr int maximum_health = 1000;
+constexpr float initial_health = 1000.0f;
+constexpr float maximum_health = 1000.0f;
 
 #endif
